Removed TrialStateChange tests and readdded GameStartEvents tests
